refactor: extract point-queue ring into PointQueue.{c,h} (no behavior change)
